requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
sqlparse==0.5.3
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
beautifulsoup4==4.14.2
billiard==4.2.2
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
h11==0.16.0
httplib2==0.31.0
idna==3.11
joblib==1.5.2
kombu==5.5.4
lxml==6.0.2
nltk==3.9.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==5.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
redis==7.0.1
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
wikipedia==1.4.0
wsproto==1.2.0
argcomplete==3.1.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
Brotli==1.1.0
catfish==4.16.4
ccsm==0.9.14.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
command-not-found==0.3
compizconfig-python==0.9.14.2
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
Deprecated==1.2.14
deprecation==2.0.7
dill==0.3.8
distro==1.9.0
eyed3==0.9.7
feedparser==6.0.10
filelock==3.13.1
filetype==1.2.0
freetype-py==2.4.0
guake==3.10
gyp==0.1
html2text==2024.2.26
httplib2==0.20.4
humanfriendly==10.0
hypothesis==6.98.15
idna==3.6
ifaddr==0.2.0
IMDbPY==2021.4.18
img2pdf==0.5.1
jupyter_core==5.3.2
kazam==1.4.5
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libevdev==0.5
libvirt-python==10.0.0
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
menulibre==2.4.0
mugshot==0.4.3
multiprocess==0.70.16
mutagen==1.46.0
mysql-connector-python==9.3.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
ocrmypdf==15.2.0+dfsg1
onboard==1.4.1
packaging==24.0
pdfminer.six==20221105
pexpect==4.9.0
pikepdf==8.7.1
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
pluggy==1.4.0
psutil==5.9.8
ptyprocess==0.7.0
pycairo==1.25.1
pycryptodomex==3.20.0
pycups==2.0.1
pycurl==7.45.3
pyelftools==0.30
Pygments==2.17.2
PyGObject==3.48.2
PyICU==2.12
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyparted==3.12.0
pypng==0.20231004.0
python-apt==2.7.7+ubuntu5
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-gnupg==0.5.2
pyudev==0.24.0
pyxdg==0.28
PyYAML==6.0.1
qrcode==7.4.2
repolib==2.2.1
reportlab==4.1.0
requests==2.31.0
requests-file==1.5.1
rich==13.7.1
rlPyCairo==0.3.0
screen-resolution-extra==0.0.0
setproctitle==1.3.3
setuptools==68.1.2
setuptools-scm==8.0.4
sgmllib3k==1.0.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
systemd-python==235
tldextract==3.1.2
traitlets==5.5.0
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ufw==0.36.2
Unidecode==1.3.8
urllib3==2.0.7
userpath==1.9.1
vboxapi==1.0
wadllib==1.3.6
websockets==10.4
wheel==0.42.0
wrapt==1.15.0
xdg==5
xkit==0.0.0
yt-dlp==2024.4.9
